requirements.txt
pyautogen
wxauto